refactored code to work with REACT Router so the HourWeatherCard can be displayed as a seperate page
