+++ weather-app/src/components/HourWeatherCard/HourWeatherCard.js
@@ -1,5 +1,3 @@
-// Place holder for hourly data presentation
-
 import React from "react";
 import "./HourWeather.css";
 import ImportAll from "../Icons/icons";
@@ -17,7 +15,6 @@ const images = ImportAll(
 
 // then need to make this appear on a seperate page when the user clicks on the DetailWeatherCard
 
-let hourlyData = "";
 export default function next24Hours({ weatherData }) {
   let now = new Date(); // Getting current date and time
   let nextDay = new Date(); // Getting the date and time 24 hours later
@@ -31,15 +28,31 @@ export default function next24Hours({ weatherData }) {
 
   console.log(hourlyData); // see below for hourlyData format
 
- 
   let compassDirection = ""; //creating a block scope variable
   // Main Program
-  function compassDirectionConversion (hourlyData) {
-   
-   const sector = ["N","NNE","NE","ENE","E","ESE","SE","SSE","S","SSW","SW","WSW","W","WNW","NW","NNW","N"]; // N can be 0deg or 360Deg so it is included twice in the array to capture this edge case.
-   let windDirection = hourlyData.values.windDirection;
-   compassDirection = sector[Math.round(windDirection / 22.5)]; // Rounds the wind direction out into 17 sectors. Sectors 0 and 16 are both N.  we could add more granularity by adding more sectors and changing the divisor according to howe many sectors we have chosen to use.
-   return compassDirection;
+  function compassDirectionConversion(hourlyData) {
+    const sector = [
+      "N",
+      "NNE",
+      "NE",
+      "ENE",
+      "E",
+      "ESE",
+      "SE",
+      "SSE",
+      "S",
+      "SSW",
+      "SW",
+      "WSW",
+      "W",
+      "WNW",
+      "NW",
+      "NNW",
+      "N",
+    ]; // N can be 0deg or 360Deg so it is included twice in the array to capture this edge case.
+    let windDirection = hourlyData.values.windDirection;
+    compassDirection = sector[Math.round(windDirection / 22.5)]; // Rounds the wind direction out into 17 sectors. Sectors 0 and 16 are both N.  we could add more granularity by adding more sectors and changing the divisor according to howe many sectors we have chosen to use.
+    return compassDirection;
   }
 
   return (
@@ -54,8 +67,6 @@ export default function next24Hours({ weatherData }) {
           <h3>Wind Gust: {hourlyData.values.windGust}</h3>
           <h3>Wind Direction: {hourlyData.values.windDirection}</h3>
           <h3>Wind Compass: {compassDirectionConversion(hourlyData)}</h3>
-          {/* convert the above to compass direction */}
-
         </div>
       ))}
     </div>
@@ -64,82 +75,34 @@ export default function next24Hours({ weatherData }) {
 
 // hourly data is an array of arrays 0 to 24
 // each array is one hour in this format:
-  // { time: '2021-03-16T12:00:00Z', 
-  //   values: {cloudBase:0.67
-             // cloudCeiling:0.71
-             // cloudCover:100
-             // dewPoint:6.44
-             // evapotranspiration:0.1
-             // freezingRainIntensity:0 
-             // humidity:69.74
-             // iceAccumulation:0
-             // iceAccumulationLwe:0
-             // precipitationProbability:0
-             // pressureSurfaceLevel:1027.32
-             // rainAccumulation:0
-             // rainAccumulationLwe:0
-             // rainIntensity:0
-             // sleetAccumulation:0
-             // sleetAccumulationLwe:0
-             // sleetIntensity:0
-             // snowAccumulation:0
-             // snowAccumulationLwe:0
-             // snowIntensity:0
-             // temperature:11.78
-             // temperatureApparent:11.78
-             // uvHealthConcern:0
-             // uvIndex:0
-             // visibility:16
-             // weatherCode:1001
-             // windDirection:57.01
-             // windGust:12.45
-             // windSpeed:7.37
-   //}
-
-
-// function HourWeatherCard({ weatherData, dayIndex }) {
-// // filter the data to get the next 24 hours
-
-//   const maxTemp = Math.round(
-//     weatherData.timelines.daily[dayIndex].values.temperatureMax
-//   );
-//   const minTemp = Math.round(
-//     weatherData.timelines.daily[dayIndex].values.temperatureMin
-//   );
-//   const avgHumidity = Math.round(
-//     weatherData.timelines.daily[dayIndex].values.humidityAvg
-//   );
-//   const probabilityRain = Math.round(
-//     weatherData.timelines.daily[dayIndex].values.precipitationProbabilityAvg
-//   );
-//   const avgWindSpeed = Math.round(
-//     weatherData.timelines.daily[dayIndex].values.windSpeedAvg * 10) / 10;
-
-//   const dateString = weatherData.timelines.daily[dayIndex].time;
-//   const dateObj = new Date(dateString);
-//   const plainDate = dateObj.toLocaleDateString(undefined, { weekday: "long" });
-//   const today = new Date().toLocaleDateString(undefined, { weekday: "long" });
-
-//   const weatherCodeMax =
-//     weatherData.timelines.daily[dayIndex].values.weatherCodeMax;
-//   // console.log(weatherCodeMax)
-//   const icon = images[`./${weatherCodeMax}.png`];
-
-//   return (
-//     <div className="DetailWeatherCard">
-//       <p className="dayHeading">{plainDate === today ? "Today" : plainDate}</p>
-//       <div className="iconAndTemp">
-//         <img src={icon} alt="weather icon" className="iconImage" />
-//         <div className="tempContainer">
-//           <p className="temp">Max Temp: {maxTemp}&deg;C</p>
-//           <p className="temp">Min Temp: {minTemp}&deg;C</p>
-//           <p className="temp">Humidity: {avgHumidity}%</p>
-//           <p className="temp">Rain: {probabilityRain}%</p>
-//           <p className="temp">Wind: {avgWindSpeed} mph</p>
-//         </div>
-//       </div>
-//     </div>
-//   );
-// }
-//
-// export default HourWeatherCard;
+// { time: '2021-03-16T12:00:00Z',
+//   values: {cloudBase:0.67
+// cloudCeiling:0.71
+// cloudCover:100
+// dewPoint:6.44
+// evapotranspiration:0.1
+// freezingRainIntensity:0
+// humidity:69.74
+// iceAccumulation:0
+// iceAccumulationLwe:0
+// precipitationProbability:0
+// pressureSurfaceLevel:1027.32
+// rainAccumulation:0
+// rainAccumulationLwe:0
+// rainIntensity:0
+// sleetAccumulation:0
+// sleetAccumulationLwe:0
+// sleetIntensity:0
+// snowAccumulation:0
+// snowAccumulationLwe:0
+// snowIntensity:0
+// temperature:11.78
+// temperatureApparent:11.78
+// uvHealthConcern:0
+// uvIndex:0
+// visibility:16
+// weatherCode:1001
+// windDirection:57.01
+// windGust:12.45
+// windSpeed:7.37
+//}
